refactor(editor): move suggestions plugin into package
